+++ core/event-service/src/main/java/ru/practicum/event/event/service/EventServiceImpl.java
@@ -1,19 +1,10 @@
 package ru.practicum.event.event.service;
 
-import com.querydsl.core.BooleanBuilder;
-import com.querydsl.core.types.dsl.BooleanExpression;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-import ru.practicum.event.category.model.Category;
-import ru.practicum.event.category.repository.CategoryRepository;
-import ru.practicum.event.event.EventRepository;
-import ru.practicum.event.event.LocationRepository;
-import ru.practicum.event.event.MapperEvent;
 import ru.practicum.event.event.model.Event;
-import ru.practicum.event.event.model.QEvent;
 import ru.practicum.event.event.service.param.GetEventAdminParam;
 import ru.practicum.event.event.service.param.GetEventUserParam;
 import ru.practicum.event.event.util.ResponseEventBuilder;
@@ -21,7 +12,6 @@ import ru.practicum.interaction.dto.event.EventFullDto;
 import ru.practicum.interaction.dto.event.EventRequestStatusUpdateRequest;
 import ru.practicum.interaction.dto.event.EventRequestStatusUpdateResult;
 import ru.practicum.interaction.dto.event.EventShortDto;
-import ru.practicum.interaction.dto.event.EventState;
 import ru.practicum.interaction.dto.event.NewEventDto;
 import ru.practicum.interaction.dto.event.UpdateEventAdminRequest;
 import ru.practicum.interaction.dto.event.UpdateEventUserRequest;
@@ -33,91 +23,36 @@ import ru.practicum.interaction.exception.NotFoundException;
 import ru.practicum.interaction.feign.request.RequestInternalFeign;
 import ru.practicum.interaction.feign.user.UserInternalFeign;
 
-import java.time.LocalDateTime;
 import java.util.List;
-import java.util.Objects;
 
 import static ru.practicum.event.event.util.ValidatorEventTime.isEventTimeBad;
-import static ru.practicum.interaction.Constants.CATEGORY_NOT_FOUND;
 import static ru.practicum.interaction.Constants.EVENT_NOT_FOUND;
-import static ru.practicum.interaction.dto.event.EventState.CANCELED;
-import static ru.practicum.interaction.dto.event.EventState.PENDING;
 import static ru.practicum.interaction.dto.event.EventState.PUBLISHED;
-import static ru.practicum.interaction.dto.event.EventState.REJECTED;
-import static ru.practicum.interaction.dto.event.UpdateEventAdminRequest.StateAction.PUBLISH_EVENT;
-import static ru.practicum.interaction.dto.event.UpdateEventUserRequest.StateAction.SEND_TO_REVIEW;
 
+/**
+ * Сервис не работает напрямую с репозиторием и не содержит транзакции (@Transactional)
+ * Для взаимодействия с доменной сущностью он использует EventTransactionService
+ * Работа с внешними сервисами происходит через клиенты Feign
+ * Благодаря разделению сокращается длительность и количество соединений с БД
+ */
 @Slf4j
 @Service
 @RequiredArgsConstructor
 public class EventServiceImpl implements EventService {
-    private final EventRepository eventRepository;
-    private final LocationRepository locationRepository;
-    private final MapperEvent eventMapper;
+    private final EventTransactionService eventTransactionService;
     private final RequestInternalFeign requestInternalFeign;
     private final UserInternalFeign userInternalFeign;
-    private final CategoryRepository categoryRepository;
     private final ResponseEventBuilder responseEventBuilder;
 
-    @Transactional(readOnly = true)
     @Override
     public List<EventFullDto> getEventsByAdmin(GetEventAdminParam param) {
-        QEvent event = QEvent.event;
-        BooleanBuilder requestBuilder = new BooleanBuilder();
-        if (param.hasUsers()) {
-            requestBuilder.and(event.initiatorId.in(param.getUsers()));
-        }
-
-        if (param.hasStates()) {
-            requestBuilder.and(event.state.in(param.getStates()));
-        }
-
-        if (param.hasCategories()) {
-            requestBuilder.and(event.category.id.in(param.getCategories()));
-        }
-
-        if (param.hasRangeStart()) {
-            requestBuilder.and(event.createdOn.gt(param.getRangeStart()));
-        }
-
-        if (param.hasRangeEnd()) {
-            requestBuilder.and(event.createdOn.lt(param.getRangeEnd()));
-        }
-
-        List<Event> events = eventRepository.findAll(requestBuilder, param.getPage()).getContent();
+        List<Event> events = eventTransactionService.getEventsByAdmin(param);
         return responseEventBuilder.buildManyEventResponseDto(events, EventFullDto.class);
     }
 
-    @Transactional(readOnly = true)
     @Override
     public List<EventShortDto> getEventsByUser(GetEventUserParam param) {
-        QEvent event = QEvent.event;
-
-        BooleanBuilder requestBuilder = new BooleanBuilder();
-
-        requestBuilder.and(event.state.eq(PUBLISHED));
-
-        if (param.hasText()) {
-            BooleanExpression descriptionExpression = event.description.like(param.getText());
-            BooleanExpression annotationExpression = event.annotation.like(param.getText());
-            requestBuilder.andAnyOf(descriptionExpression, annotationExpression);
-        }
-
-        if (param.hasCategories()) {
-            requestBuilder.and(event.category.id.in(param.getCategories()));
-        }
-
-        if (param.hasPaid()) {
-            requestBuilder.and(event.paid.eq(param.getPaid()));
-        }
-
-        requestBuilder.and(event.eventDate.gt(Objects.requireNonNullElseGet(param.getRangeStart(), LocalDateTime::now)));
-
-        if (param.hasRangeEnd()) {
-            requestBuilder.and(event.eventDate.lt(param.getRangeEnd()));
-        }
-
-        List<Event> events = eventRepository.findAll(requestBuilder, param.getPage()).getContent();
+        List<Event> events = eventTransactionService.getEventsByUser(param);
         List<EventShortDto> eventDtos = responseEventBuilder.buildManyEventResponseDto(events, EventShortDto.class);
 
         if (param.getOnlyAvailable()) {
@@ -127,10 +62,9 @@ public class EventServiceImpl implements EventService {
         return eventDtos;
     }
 
-    @Transactional(readOnly = true)
     @Override
     public List<EventShortDto> getAllUsersEvents(Long userId, Pageable page) {
-        List<Event> events = eventRepository.findByInitiatorId(userId, page);
+        List<Event> events = eventTransactionService.getAllUsersEvents(userId, page);
         return responseEventBuilder.buildManyEventResponseDto(events, EventShortDto.class);
     }
 
@@ -140,53 +74,20 @@ public class EventServiceImpl implements EventService {
             throw new BadRequestException("Дата и время на которые намечено событие не может быть раньше, чем через два часа от текущего момента");
         }
         userInternalFeign.findUserById(userId);
-        Event event = addNewEventInTransaction(eventDto, userId);
+        Event event = eventTransactionService.createEvent(eventDto, userId);
 
-        return responseEventBuilder.buildOneEventResponseDto(event, EventFullDto.class);
-    }
-
-    @Transactional(readOnly = true)
-    @Override
-    public EventFullDto getEventForUser(Long userId, Long eventId) {
-        Event event = eventRepository.findByIdAndInitiatorId(eventId, userId)
-                .orElseThrow(() -> new NotFoundException(EVENT_NOT_FOUND));
         return responseEventBuilder.buildOneEventResponseDto(event, EventFullDto.class);
     }
 
     @Override
-    @Transactional
+    public EventFullDto getEventForUser(Long userId, Long eventId) {
+        Event event = eventTransactionService.getEventForUser(userId, eventId);
+        return responseEventBuilder.buildOneEventResponseDto(event, EventFullDto.class);
+    }
+
+    @Override
     public EventFullDto updateEventByUser(Long userId, Long eventId, UpdateEventUserRequest updateDto) {
-        Event event = eventRepository.findByIdAndInitiatorId(eventId, userId)
-                .orElseThrow(() -> new NotFoundException(EVENT_NOT_FOUND));
-
-        EventState state = event.getState();
-        if (state == PUBLISHED) {
-            throw new ConflictException("Изменить можно только не опубликованные события, текущий статус " + state);
-        }
-
-        if (updateDto.hasStateAction()) {
-            if (updateDto.getStateAction().equals(SEND_TO_REVIEW)) {
-                event.setState(PENDING);
-            } else {
-                event.setState(CANCELED);
-            }
-        }
-
-        if (updateDto.hasEventDate()) {
-            if (isEventTimeBad(updateDto.getEventDate(), 2)) {
-                throw new BadRequestException("Дата начала изменяемого события должна быть не ранее чем за 2 часа от даты публикации");
-            }
-            event.setEventDate(updateDto.getEventDate());
-        }
-
-        if (updateDto.hasCategory()) {
-            Category category = categoryRepository.findById(updateDto.getCategory())
-                    .orElseThrow(() -> new NotFoundException(CATEGORY_NOT_FOUND));
-            event.setCategory(category);
-        }
-
-        eventMapper.updateEventByUserRequest(event, updateDto);
-
+        Event event = eventTransactionService.updateEventByUser(userId, eventId, updateDto);
         return responseEventBuilder.buildOneEventResponseDto(event, EventFullDto.class);
     }
 
@@ -202,8 +103,7 @@ public class EventServiceImpl implements EventService {
                                                               EventRequestStatusUpdateRequest updateRequest) {
         userInternalFeign.findUserShortById(userId);
 
-        Event event = eventRepository.findByIdAndInitiatorId(eventId, userId)
-                .orElseThrow(() -> new NotFoundException(EVENT_NOT_FOUND));
+        Event event = eventTransactionService.getEventForUser(userId, eventId);
 
         if (isPreModerationOff(event.getRequestModeration(), event.getParticipantLimit())) {
             return new EventRequestStatusUpdateResult();
@@ -217,13 +117,11 @@ public class EventServiceImpl implements EventService {
         return requestInternalFeign.updateEventRequests(patchDto);
     }
 
-    @Transactional(readOnly = true)
     @Override
     public EventFullDto getEventById(Long eventId, boolean isInternalRequest) {
-        Event eventDomain = eventRepository.findById(eventId)
-                .orElseThrow(() -> new NotFoundException(EVENT_NOT_FOUND));
+        Event event = eventTransactionService.getEventById(eventId);
 
-        if (eventDomain.getState() != PUBLISHED) {
+        if (event.getState() != PUBLISHED) {
             if (isInternalRequest) {
                 throw new ConflictException("Событие ещё не опубликовано.");
             } else {
@@ -231,57 +129,13 @@ public class EventServiceImpl implements EventService {
             }
         }
 
-        return responseEventBuilder.buildOneEventResponseDto(eventDomain, EventFullDto.class);
-    }
-
-    @Override
-    @Transactional
-    public EventFullDto updateEventByAdmin(Long eventId, UpdateEventAdminRequest updateDto) {
-        Event event = eventRepository.findById(eventId)
-                .orElseThrow(() -> new NotFoundException(EVENT_NOT_FOUND));
-
-        if (event.getState() != EventState.PENDING) {
-            throw new ConflictException("Изменить можно только события ожидающие модерацию, текущий статус " + event.getState());
-        }
-
-        if (updateDto.hasStateAction()) {
-            if (updateDto.getStateAction() == PUBLISH_EVENT) {
-                event.setState(PUBLISHED);
-                event.setPublishedOn(LocalDateTime.now());
-            } else {
-                event.setState(REJECTED);
-            }
-        }
-
-        if (updateDto.hasEventDate()) {
-            if (isEventTimeBad(updateDto.getEventDate(), 1)) {
-                throw new BadRequestException("Дата начала изменяемого события должна быть не ранее чем за час от даты публикации");
-            }
-            event.setEventDate(updateDto.getEventDate());
-        }
-
-        if (updateDto.hasCategory()) {
-            Category category = categoryRepository.findById(updateDto.getCategory())
-                    .orElseThrow(() -> new NotFoundException(CATEGORY_NOT_FOUND));
-            event.setCategory(category);
-        }
-
-        eventMapper.updateEventByAdminRequest(event, updateDto);
-
         return responseEventBuilder.buildOneEventResponseDto(event, EventFullDto.class);
     }
 
-    @Transactional
-    private Event addNewEventInTransaction(NewEventDto eventDto, Long userId) {
-        Category category = categoryRepository.findById(eventDto.getCategory()).orElseThrow(
-                () -> new NotFoundException(CATEGORY_NOT_FOUND));
-
-        Event eventFromDb = eventMapper.toEvent(eventDto, category, userId);
-        eventFromDb.getLocation().setEvent(eventFromDb);
-        locationRepository.save(eventFromDb.getLocation());
-
-        eventRepository.save(eventFromDb);
-        return eventFromDb;
+    @Override
+    public EventFullDto updateEventByAdmin(Long eventId, UpdateEventAdminRequest updateDto) {
+        Event event = eventTransactionService.updateEventByAdmin(eventId, updateDto);
+        return responseEventBuilder.buildOneEventResponseDto(event, EventFullDto.class);
     }
 
     private boolean isPreModerationOff(boolean moderationStatus, int limit) {
